Define start and stop commands on the CLI
I'm not entirely sure what the UI for this application looks like yet,
but I do know that we're going to need a way of starting a session and
stopping one. My initial ideas are something like the following:

        # Start a new session, this will error with a friendly message if
        # there's already an active session
        time_tracker start "Refactor the Foo Class"
        # This will stop the currently active session, is a no-op if no
        # session is running
        time_tracker stop

There are a couple of things to note about this commit:

* Subcommands operates directly on ARGV

I didn't want to have to work out how to implement subcommands using
OptionParser, so I found the SubCommand gem. It seems OK, it's a single
class that wraps OptionParser so isn't too much of a maintenance burden.
However I don't like that it operates on ARGV directly, as this means
that it's less obvious that I have control over it,, and it's also much
less obvious in CLI that ARGV is being used - in fact ARGV doesn't
appear anywhere in my code at all anymore, which is a bit odd.

What I'd like is if I could collect ARGV in my binary file and then pass
that down to the CLI and explicitly pass it (in the form of an options
hash) through to Subcommand.

* Subcommands are defined using a definition hash

Instead of using the Subcommand DSL repeatedly and explicitly switching
on the output of opt_parse() I've decided to define subcommand names,
descriptions and the code that they dispatch to in a hash which is
parsed into the subcommnd definitions at instantiation time. I quite
like this approach as it's less verbose than using the DSL repeatedly
and it keeps the subcommnd definitions close to each other, and easy to
read and reason about. Hopefully this pattern will be flexible enough to
scale as the program growns, but I'm not wedded to it.