Build a consistent interface for file system storage
I need to be able to store application state in order to save session
details and maintain running sessions between runs of this application.

In order to do that I need a place to store data. This DataLocation
class attempts to give me that by papering over which Operating system
you're running and just returning the standard user data storage
location for applications relevant to your system.

* It doesn't support Windows

Maybe it might in the future, but I don't use Windows and have no way of
testing so Linux/MacOS only for now

* I'm explicitly supporting Linux rather than Posix.

I know nothing about UNIX systems that are not MacOS or Linux, so I
don't feel confident enough yet to say whether this will work on say
FreeBSD or Haiku or something like that

* Testing this is unpleasant

I started off trying to do

        ENV.stubs(:[]).with('XDG_DATA_HOME')

but for whathever reason this doesn't work because Mocha replaces the
entire :[] method, and it's called at other places deep in the stack
with arguments like 'TMP' and 'TMPDIR'. I spent some time with byebug
trying to find out whether this occurs inside Ruby, or inside Mocha or
Minitest - I couldn't fnd it so I gave up and just opted for the
approach I went for which is to replace the values in the actual ENV
hash for the duration of the tests and then reset them in the teardown.

It's not very elegant but it works.

* At some point I'll take care of error handling properly

defining BrokenConfig at the top of the class in that way is ok for now,
and I wanted to throw something that was more specific than KeyError.
but I need to be careful that I don't end up with a load of these ghetto
exception classes defined all over the codebase.