Fix Zeitwerk loading issue.
This commit fixes an intermittent code loading problem that occured from
the data location tests.

On certain seeds, the test runner runs the tests so that the
DataLocationGenericTests get run first. In this case when parsing the
file, ruby attempts to load BrokenConfig first, because it's th first
Zeitwerk controlled constant the interpreter encounters in that case.

The problem is that because BrokenConfig is no defined in it's own file
(it's defined inside data_location.rb, but as an ancestor of
TimeTracker) Zeitwerk doesn't know where to find it and therefore it's
not loaded.

I could have defined a custom autoload on the Zeitwerk loader but
introducing global config felt like too much of a sledgehammer to crack
this particular nut. Instead I've moved the constant inside the
DataLocation class and refer to it specifically in the tests to ensure
it's always loaded correctly. This makes sense as BrokenConfig is only
ever raised from the DataLocation class at the moment anyway.